feat: propagate errors as objects and align Winston service
